# Kafka 9626: Improve ACLAuthorizer.acls() performance