@omkreddy 2020-03-01
Thanks to @lbradstreet for noticing the issue and suggesting the fix.