@lbradstreet 2020-03-02
@omkreddy thanks for the PR. LGTM too. Thanks for adding a benchmark!